scripts/requirements.txt
xoa-driver>=2.9.3
matplotlib>=3.9.2
pyyaml>=6.0.1
pydantic>=2.10.6